langchain/requirements.txt
langchain
langchain_groq
python-dotenv
